Slight optimization of FastSim
Changed CircuitElementValue::pressure to be a static constexpr, which
makes the divisor compile time const in the division in
CircuitElementValve::sim, allowing the compiler to turn it into a
multiplication followed by right shift. This should be bit-for-bit
equivalent to the original code, assuming the dividend fits in 32 bits.

Rearranged the expressions in FastSimPipe3 and FastSimPipe4. If you take
the sum of the arguments to a.move, a pattern emerges:
```
a_delta = -(a-b)/3 -(a-c)/3
        = (-a+b-a+c)/3
        = (-2a+b+c)/3
        = (a+b+c - 3a)/3
        = (a+b+c)/3 - a
```
This generalizes to `(a+b+c+...)/n - a` for all ports in a n-way valve.

I benchmarked the code before and after by setting the speed slider to
max and letting it settle Latch5 and Latch10. I get the following
simticks per second (eyeballing for the most common value):

```
        Before After
Latch5  51000  57000
Latch10 19700  22000
```